Move visualize_json() and generate_visualize_html() to LibASR utils
