+++ src/lfortran/utils.cpp
@@ -100,157 +100,4 @@ std::string get_runtime_library_c_header_dir()
     return path;
 }
 
-std::string generate_visualize_html(std::string &astr_data_json) {
-    std::hash<std::string> hasher;
-    std::ofstream out;
-    std::string file_name = "visualize" + std::to_string(hasher(astr_data_json)) + ".html";
-    out.open(file_name);
-    out << R"(<!DOCTYPE html>
-<html>
-<head>
-    <title>LCompilers AST/R Visualization</title>
-    <script crossorigin src="https://unpkg.com/react@18/umd/react.production.min.js"></script>
-    <script crossorigin src="https://unpkg.com/react-dom@18/umd/react-dom.production.min.js"></script>
-
-    <script src="https://unpkg.com/@babel/standalone/babel.min.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/react-flow-renderer/10.3.17/umd/index.js"></script>
-    <script src="https://dagrejs.github.io/project/dagre/latest/dagre.min.js"></script>
-    <script> )";
-    out << "var astr_data = " << astr_data_json << "; </script>\n";
-    out << R"(</head>
-
-<body style="margin: 0px;">
-    <script type="text/babel" data-type="module">
-function TreeNode({ node }) {
-    if (node.literals.length === 0) return <p><b>{node.node}</b></p>;
-    return (
-        <div>
-            <p><b>{node.node}</b></p>
-            <div style={{ backgroundColor: "#FBBD23", padding: "2px" }}>
-                {
-                    node.literals.map((val, idx) => <p style={{ margin: "0px", padding: "1px" }} key={idx}>{val[0]}: {val[1]}</p>)
-                }
-            </div>
-        </div>
-    );
-}
-
-const getLayoutedElements = (nodes, edges, direction = 'TB') => {
-    const nodeWidth = 180;
-    const isHorizontal = direction === 'LR';
-
-    const dagreGraph = new dagre.graphlib.Graph();
-    dagreGraph.setDefaultEdgeLabel(() => ({}));
-    dagreGraph.setGraph({ rankdir: direction });
-
-    nodes.forEach(node => dagreGraph.setNode(node.id, { width: nodeWidth, height: node.nodeHeight }));
-    edges.forEach(edge => dagreGraph.setEdge(edge.source, edge.target));
-
-    dagre.layout(dagreGraph);
-
-    nodes.forEach((node) => {
-        const nodeWithPosition = dagreGraph.node(node.id);
-        node.targetPosition = isHorizontal ? 'left' : 'top';
-        node.sourcePosition = isHorizontal ? 'right' : 'bottom';
-        // Shifting the dagre node position (anchor=center center) to the top left
-        // so it matches the React Flow node anchor point (top left).
-        node.position = {
-            x: nodeWithPosition.x - nodeWidth / 2,
-            y: nodeWithPosition.y - node.nodeHeight / 2,
-        };
-        return node;
-    });
-
-    return [nodes, edges];
-};
-
-class Graph {
-    constructor() {
-        this.nodes = [];
-        this.edges = [];
-        this.idx = 1;
-        return this;
-    }
-
-    createNode(cur_node) {
-        cur_node.idx = this.idx++;
-        cur_node.literals = [];
-        let obj = cur_node.fields;
-        for (let prop in obj) {
-            let neigh = obj[prop];
-            if (typeof neigh === 'object') {
-                if (neigh.hasOwnProperty("node")) {
-                    this.createEdge(cur_node.idx, neigh, prop);
-                } else {
-                    if (neigh.length > 0) {
-                        for (let i in neigh) {
-                            let arrayElement = neigh[i];
-                            if (typeof arrayElement === 'object') {
-                                if (arrayElement.hasOwnProperty("node")) {
-                                    this.createEdge(cur_node.idx, arrayElement, `${prop}[${i}]`);
-                                } else {
-                                    console.log("ERROR: Unexpected 2D Array found");
-                                }
-                            } else {
-                                cur_node.literals.push([`${prop}[${i}]`, `${arrayElement}`]);
-                            }
-                        }
-                    } else {
-                        // 0 length array, show as literal
-                        cur_node.literals.push([prop, "[]"]);
-                    }
-                }
-            } else {
-                cur_node.literals.push([prop, `${neigh}`]);
-            }
-        }
-
-        this.nodes.push({ id: `${cur_node.idx}`, data: { label: <TreeNode node={cur_node} /> }, nodeHeight: 70 + 20 * (cur_node.literals.length) });
-    }
-
-    createEdge(parent_idx, cur_node, edge_label) {
-        this.edges.push({
-            id: `${parent_idx}-${this.idx}`,
-            source: `${parent_idx}`,
-            target: `${this.idx}`,
-            label: edge_label,
-            labelStyle: { fontWeight: 700 },
-            labelBgPadding: [8, 4],
-            labelBgStyle: { fill: '#FBBD23' },
-        });
-        this.createNode(cur_node);
-    }
-}
-
-function Flow({ nodes, edges }) {
-    return (
-        <div style={{ height: '100vh' }}>
-            <ReactFlow.default
-                defaultNodes={nodes}
-                defaultEdges={edges}
-                style={{ backgroundColor: '#e5e7eb' }}
-            >
-                <ReactFlow.Background />
-                <ReactFlow.Controls />
-                <ReactFlow.MiniMap />
-            </ReactFlow.default>
-        </div>
-    );
-}
-
-function MyApp() {
-    var g = new Graph();
-    g.createNode(astr_data);
-    var [layoutedNodes, layoutedEdges] = getLayoutedElements(g.nodes, g.edges);
-    return (<Flow nodes={layoutedNodes} edges={layoutedEdges} />);
-}
-
-ReactDOM.render(<MyApp />, document.body);
-    </script>
-</body>
-
-</html>)";
-    return file_name;
-}
-
 } // namespace LCompilers::LFortran
