+++ src/lfortran/utils.h
@@ -10,7 +10,6 @@ void get_executable_path(std::string &executable_path, int &dirname_length);
 std::string get_runtime_library_dir();
 std::string get_runtime_library_header_dir();
 std::string get_runtime_library_c_header_dir();
-std::string generate_visualize_html(std::string &astr_data_json);
 
 } // LCompilers::LFortran
